fix: schedule input type 구분

+++ calendar-service/src/main/java/com/example/calendarservice/dto/RepeatScheduleRequest.java
@@ -9,8 +9,8 @@ import java.util.List;
 
 @Data
 public class RepeatScheduleRequest {
-    private Integer branchId;
-    private Integer userId;
+    private Long branchId;
+    private Long userId;
     private LocalTime startTime;
     private LocalTime endTime;
     private Repeat repeat;
